backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
anthropic
sqlalchemy
pydantic
sqlalchemy-utils
psycopg2-binary
alembic
python-jose
passlib
bcrypt
